Renaming the entire package properly to complement the module rename
