+++ phantom-streams/src/test/scala/com/outworkers/phantom/streams/suites/PublisherIntegrationTest.scala
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.outworkers.phantom.reactivestreams.suites
+package com.outworkers.phantom.streams.suites
 
 import java.util.concurrent.atomic.AtomicInteger
 
 import com.outworkers.phantom.dsl._
-import com.outworkers.phantom.reactivestreams._
+import com.outworkers.phantom.streams._
 import com.outworkers.util.testing._
 import org.reactivestreams.{Subscriber, Subscription}
 import org.scalatest.FlatSpec
