+++ phantom-streams/src/test/scala/com/outworkers/phantom/streams/suites/iteratee/IterateePerformanceTest.scala
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.outworkers.phantom.reactivestreams.suites.iteratee
+package com.outworkers.phantom.streams.suites.iteratee
 
 import java.util.concurrent.atomic.AtomicInteger
 
 import com.outworkers.phantom.PhantomSuite
 import com.outworkers.phantom.dsl._
-import com.outworkers.phantom.reactivestreams._
+import com.outworkers.phantom.streams._
 import com.outworkers.phantom.tables._
 import com.outworkers.util.testing._
 import org.scalatest.concurrent.PatienceConfiguration
