+++ phantom-streams/src/test/scala/com/outworkers/phantom/streams/suites/iteratee/OperaPublisher.scala
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.outworkers.phantom.reactivestreams.suites.iteratee
+package com.outworkers.phantom.streams.suites.iteratee
 
 import java.util.concurrent.atomic.AtomicInteger
 
-import com.outworkers.phantom.reactivestreams.suites.{Opera, OperaData}
+import com.outworkers.phantom.streams.suites.{Opera, OperaData}
 import org.reactivestreams.{Publisher, Subscriber, Subscription}
 
 object OperaPublisher extends Publisher[Opera] {
